+++ packages/components-dev/all/module.ts
@@ -5,11 +5,11 @@ import { KbqMomentDateModule } from '@koobiq/angular-moment-adapter/adapter';
 import { KbqButtonModule } from '@koobiq/components/button';
 import { KbqButtonToggleModule } from '@koobiq/components/button-toggle';
 import { KbqCheckboxModule } from '@koobiq/components/checkbox';
-import { PopUpPlacements, ThemePalette } from '@koobiq/components/core';
+import { KbqNormalizeWhitespace, PopUpPlacements, ThemePalette } from '@koobiq/components/core';
 import { KbqDropdownModule } from '@koobiq/components/dropdown';
 import { KbqFormFieldModule } from '@koobiq/components/form-field';
 import { KbqIconModule } from '@koobiq/components/icon';
-import { KbqInputModule, KbqNormalizeWhitespace } from '@koobiq/components/input';
+import { KbqInputModule } from '@koobiq/components/input';
 import { KbqLinkModule } from '@koobiq/components/link';
 import { KbqListModule } from '@koobiq/components/list';
 import { KbqMarkdownModule } from '@koobiq/components/markdown';
