+++ packages/components/core/common-behaviors/index.ts
@@ -1,6 +1,7 @@
 import { InjectionToken } from '@angular/core';
 
 export * from './checkbox';
+export * from './clipboard';
 export { CanColor, CanColorCtor, KbqColorDirective, KbqComponentColors, mixinColor, ThemePalette } from './color';
 export { KBQ_SANITY_CHECKS, KbqCommonModule } from './common-module';
 export { CanDisable, CanDisableCtor, mixinDisabled } from './disabled';
