+++ packages/components/input/input-number.ts
@@ -6,7 +6,6 @@ import {
     ElementRef,
     EventEmitter,
     forwardRef,
-    inject,
     Inject,
     Input,
     OnDestroy,
@@ -89,31 +88,6 @@ export const KBQ_NUMBER_INPUT_VALUE_ACCESSOR: any = {
     useExisting: forwardRef(() => KbqNumberInput),
     multi: true
 };
-
-@Directive({
-    standalone: true,
-    selector: '[kbqNormalizeWhitespace]',
-    host: {
-        '(copy)': 'onCopy($event)'
-    }
-})
-export class KbqNormalizeWhitespace {
-    /** @docs-private */
-    protected elementRef: ElementRef<HTMLElement> = inject(ElementRef);
-
-    /**
-     * Replace thin-space with space on copy event
-     */
-    onCopy($event: ClipboardEvent) {
-        const value =
-            (this.elementRef.nativeElement as HTMLInputElement).value || this.elementRef.nativeElement.textContent;
-
-        if ($event.type === 'copy' && value) {
-            $event.preventDefault();
-            $event.clipboardData?.setData('text', value?.replace(/\u2009/g, ' '));
-        }
-    }
-}
 
 @Directive({
     selector: `input[kbqNumberInput]`,
