+++ packages/docs-examples/components/input/input-number-overview/input-number-overview-example.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { KbqFormsModule } from '@koobiq/components/core';
+import { KbqFormsModule, KbqNormalizeWhitespace } from '@koobiq/components/core';
 import { KbqFormFieldModule } from '@koobiq/components/form-field';
-import { KbqInputModule, KbqNormalizeWhitespace } from '@koobiq/components/input';
+import { KbqInputModule } from '@koobiq/components/input';
 
 /**
  * @title Input number
